+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Line wrap the file at 100 chars.                                              Th
 
 ### Fixed
 - Mark CLI `bridge set state` argument as required to avoid a crash.
+- The VPN service on Windows will now be restarted when it crashes.
 
 
 ## [2019.6] - 2019-07-15